Uppdated GUI
Reworked the way it generates the boxes so that it can draw differnt
objects.

The GUI is now a state machine that draw whatever its state is equal
to. An example of the switch changes can be seen with the in program
menu.

The text rendering shader has been slightly modififed to ensure that
text apears on top of the boxs.